+++ src/i18n/i18n.ts
@@ -23,6 +23,7 @@ const {
     logoutConfirmationModal: enLogoutConfirmationModal,
     error: enError,
     bookingPageConfirmation: enBookingPageConfirmation,
+    errorPage: enErrorPage,
 } = en;
 
 const {
@@ -45,6 +46,7 @@ const {
     logoutConfirmationModal: esLogoutConfirmationModal,
     error: esError,
     bookingPageConfirmation: esBookingPageConfirmation,
+    errorPage: esErrorPage,
 } = es;
 
 export const defaultNS = 'ns1';
@@ -72,6 +74,7 @@ void i18n.use(initReactI18next).init({
             logoutConfirmationModal: enLogoutConfirmationModal,
             error: enError,
             bookingPageConfirmation: enBookingPageConfirmation,
+            errorPage: enErrorPage,
         },
         es: {
             ns1: esNs1,
@@ -93,6 +96,7 @@ void i18n.use(initReactI18next).init({
             logoutConfirmationModal: esLogoutConfirmationModal,
             error: esError,
             bookingPageConfirmation: esBookingPageConfirmation,
+            errorPage: esErrorPage,
         },
     },
     lng: language,
@@ -118,6 +122,7 @@ void i18n.use(initReactI18next).init({
         'tableExportOptions',
         'logoutConfirmationModal',
         'bookingPageConfirmation',
+        'errorPage',
     ],
     interpolation: {
         escapeValue: false, //escape dynamic content and opting not to have the i18n library perform additional escaping for the interpolated values.